splitting sidebar item helper
